# src/config/swagger.yaml
openapi: 3.0.0
info:
  title: Admin Management API
  description: API documentation for managing admin resources
  version: 1.0.0
servers:
  - url: https://backend-tour-booking-node-js-mongodb.onrender.com/api/v1/admin
    description: Admin use router this management
    variables:
      environment:
        default: production
  - url: https://backend-tour-booking-node-js-mongodb.onrender.com/api/v1/auth
    description: Traveler and Guide use this
    variables:
      environment:
        default: production
paths:
...
  /send-otp:
    post:
      summary: Send OTP for password recovery
      operationId: sendOtp
      tags:
        - Authentication
      responses:
        "200":
          description: OTP sent to the user's email
        "400":
          description: Bad request
  /new-password:
    post:
      summary: Set a new password
      operationId: newPassword
      tags:
        - Authentication
      requestBody:
        description: New password and OTP
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                otp:
                  type: string
                newPassword:
                  type: string
      responses:
        "200":
          description: Password updated successfully
        "400":
          description: Invalid OTP or password
  /profile:
    get:
      summary: Get user profile
      operationId: getProfile
      tags:
        - Authentication
      responses:
        "200":
          description: User profile details
        "401":
          description: Unauthorized
  /post-blog:
    post:
      summary: Create a new blog post
      operationId: postBlog
      tags:
        - Authentication
      requestBody:
        description: Blog post details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        "201":
          description: Blog post created
        "400":
          description: Bad request
  /my-blogs:
    get:
      summary: Get all blog posts by the user
      operationId: getAllBlog
      tags:
        - Authentication
      responses:
        "200":
          description: List of user's blog posts
        "401":
          description: Unauthorized
  /blog/{id}:
    get:
      summary: Get blog details by ID
      operationId: detailBlog
      tags:
        - Authentication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post details
        "404":
          description: Blog not found
  /edit-blog/{id}:
    put:
      summary: Edit a blog post by ID
      operationId: editBlog
      tags:
        - Authentication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        description: Updated blog post details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        "200":
          description: Blog post updated
        "404":
          description: Blog not found
  /remove-blog/{id}:
    delete:
      summary: Remove a blog post by ID
      operationId: removeBlog
      tags:
        - Authentication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post deleted
        "404":
          description: Blog not found
  /comment-blog/{id}:
    post:
      summary: Comment on a blog post
      operationId: commentBlog
      tags:
        - Authentication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        description: Comment details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
      responses:
        "201":
          description: Comment added
        "404":
          description: Blog not found
  /edit-comment/{id}:
    patch:
      summary: Edit a comment
      operationId: updateComment
      tags:
        - Authentication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        description: Updated comment details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
      responses:
        "200":
          description: Comment updated
        "404":
          description: Comment or blog not found
  /remove-comment/{id}:
    delete:
      summary: Remove a comment
      operationId: removeComment
      tags:
        - Authentication
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comment removed
        "404":
          description: Comment or blog not found
  /book-tour/{tourId}:
    post:
      summary: Book a tour and proceed to payment
      operationId: tourPayment
      tags:
        - Authentication
      parameters:
        - in: path
          name: tourId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tour booked successfully
        "400":
          description: Bad request
  /my-book-tour:
    get:
      summary: Get list of tours booked by the user
      operationId: getBookedList
      tags:
        - Authentication
      responses:
        "200":
          description: List of booked tours
        "401":
          description: Unauthorized
  /evaluate-guide/{guideId}:
    post:
      summary: Evaluate a tour guide
      operationId: evaluateGuide
      tags:
        - Authentication
      parameters:
        - in: path
          name: guideId
          required: true
          schema:
            type: string
      requestBody:
        description: Guide evaluation details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        "200":
          description: Guide evaluated
        "400":
          description: Bad request
        "404":
          description: Guide not found